修改后台界面

+++ jpkc/src/com/jpkc/web/controller/NoticeController.java
@@ -1,9 +1,12 @@
 package com.jpkc.web.controller;
 
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.List;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -11,6 +14,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.jpkc.commons.Page;
 import com.jpkc.commons.Render;
@@ -29,12 +33,12 @@ import com.jpkc.util.VerifyUtil;
 @Controller
 @RequestMapping("/notice")
 public class NoticeController {
-	
+
 	private static Log log = LogFactory.getLog(NoticeController.class);
 
 	@Resource
 	private NoticeService noticeService;
-	
+
 	/**
 	 * 添加公告
 	 * 
@@ -43,10 +47,10 @@ public class NoticeController {
 	 */
 	@RequestMapping("/add")
 	public String addNoticeIndex() {
-		
+
 		return "console/notice/add";
 	}
-	
+
 	/**
 	 * 公告列表
 	 * 
@@ -54,20 +58,20 @@ public class NoticeController {
 	 * @2015-11-8
 	 */
 	@RequestMapping("/list")
-	public String noticeList(Model model,String title) {
+	public String noticeList(Model model, String title) {
 		Notice notice = new Notice();
-		if(!VerifyUtil.isEmpty(title)) {
+		if (!VerifyUtil.isEmpty(title)) {
 			notice.setTitle(title);
 		}
-		int pageNumber = 1;
-		int pageSize = 3;
-		Page<Notice> page = noticeService.getNoticeByPage(notice,pageSize,pageNumber);
+		int pageNumber = 0;
+		int pageSize = 0;
+		Page<Notice> page = noticeService.getNoticeByPage(notice, pageSize, pageNumber);
 		model.addAttribute("page", page);
 		model.addAttribute("title", title);
-		
+
 		return "console/notice/noticelist";
 	}
-	
+
 	/**
 	 * 公告列表
 	 * 
@@ -75,13 +79,13 @@ public class NoticeController {
 	 * @2015-11-8
 	 */
 	@RequestMapping("/list/{pageNumber}/{pageSize}")
-	public String noticeList(Model model,String content,@PathVariable int pageNumber, @PathVariable int pageSize ) {
+	public String noticeList(Model model, String content, @PathVariable int pageNumber, @PathVariable int pageSize) {
 		Notice notice = new Notice();
-		Page<Notice> page = noticeService.getNoticeByPage(notice,pageSize,pageNumber);
+		Page<Notice> page = noticeService.getNoticeByPage(notice, pageSize, pageNumber);
 		model.addAttribute("page", page);
 		return "console/notice/noticelist";
 	}
-	
+
 	/***
 	 * 公告API
 	 * 
@@ -92,15 +96,22 @@ public class NoticeController {
 	 * @return
 	 */
 	@RequestMapping("/noticeList")
-	public Render<Object> noticeList(int pageNumber, int pageSize, HttpServletRequest request, Model model) {
+	public void noticeList(int pageNumber, int pageSize, HttpServletRequest request, HttpServletResponse response) {
 		Render<Object> render = new Render<Object>();
 		Notice notice = new Notice();
-		Page<Notice> page = noticeService.getNoticeByPage(notice,pageSize,pageNumber);
+		Page<Notice> page = noticeService.getNoticeByPage(notice, pageSize, pageNumber);
 		render.setData(page);
 		render.setCode("2000");
-		return render;
+		PrintWriter out;
+		try {
+			out = response.getWriter();
+			out.write(render.toString());
+			out.flush();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 	}
-	
+
 	/**
 	 * 添加公告
 	 * 
@@ -116,21 +127,31 @@ public class NoticeController {
 		noticeService.add(notice);
 		return "redirect:/notice/list";
 	}
-	
+
 	@RequestMapping("/edit/{id}")
-	public String editNotice(@PathVariable int id,Model model) {
+	public String editNotice(@PathVariable int id, Model model) {
 		Notice notice = new Notice();
 		notice.setId(id);
 		List<Notice> noticeList = noticeService.select(notice);
 		model.addAttribute("noticeList", noticeList);
 		return "console/notice/add";
 	}
-	
-	@RequestMapping("/del/{id}")
-	public String delNotice(@PathVariable int id) {
-		Notice notice = new Notice();
-		notice.setId(id);
-		noticeService.delete(notice);
+
+	@RequestMapping("/del/{ids}")
+	public String delNotice(@PathVariable int[] ids, RedirectAttributes ra) {
+		Render<Object> render = new Render<Object>();
+		if (VerifyUtil.isEmpty(ids)) {
+			log.info("删除资源信息的id是空!");
+			render.setCode("20000");
+			render.setData("删除资源id是空!");
+			ra.addFlashAttribute("render", render);
+			return "redirect:/notice/list";
+		}
+
+		for (int i = 0; i < ids.length; i++) {
+			Notice notice = new Notice(ids[i]);
+			noticeService.delete(notice);
+		}
 		return "redirect:/notice/list";
 	}
 }
